refactor: split cmd/power-manage-agent/main.go by subcommand (#66)
cmd/power-manage-agent/main.go was 1879 lines and packed the agent
daemon entrypoint, cert-rotation goroutine, periodic-sync loop,
security-alert flusher, and every CLI subcommand (setup, query,
luks, enroll, tty, self-test, version) into one file. Every change
to one subcommand recompiled the whole file and forced reviewers
to scroll past 14 unrelated functions to find the diff.

Per the proposal in agent #66, split into 11 files in the same
`main` package — pure file moves, no behaviour changes:

  main.go           — main(), parseFlags, version + constants  (352 LOC)
  runtime.go        — runAgent + sync helpers                  (368 LOC)
  cert_rotation.go  — startCertRotation + register             (203 LOC)
  backend.go        — randomBackoff + applyBackendOverrides
                       + normalizedServiceBackend              (140 LOC)
  luks_store.go     — clientLuksKeyStore                       (21 LOC)
  cmd_setup.go      — runSetup                                 (50 LOC)
  cmd_query.go      — runQuery + print helpers                 (203 LOC)
  cmd_luks.go       — runLuks + runLuksURI + runLuksSetPassphrase (261 LOC)
  cmd_enroll.go     — runEnroll + parseRegistrationURI
                       + trySocketEnroll + supporting types    (173 LOC)
  cmd_selftest.go   — runSelfTest                              (116 LOC)
  cmd_tty.go        — runTTY + printTTYUsage                   (113 LOC)

Notes:

* Same `main` package — no API changes, no exports added.
* `cmd_selftest.go` deliberately avoids the `_test.go` suffix —
  Go's test discovery treats `*_test.go` files as test files,
  which would have made `runSelfTest` invisible to the production
  build (caught during initial verification).
* `registrationURI` type and `unixSocketHTTPClient` helper moved
  alongside their only callers (cmd_enroll.go) for cohesion.
* All tests still pass (`GOWORK=off go test ./...` clean across
  every internal/* package and the new cmd/ layout).
* `goimports -w` ran on every new file to prune the import set
  to exactly what each file uses.
