+++ cmd/power-manage-agent/luks_store.go
@@ -1,0 +1,21 @@
+// Package main is the entry point for the power-manage agent.
+package main
+
+import (
+	"context"
+
+	sdk "github.com/manchtools/power-manage/sdk/go"
+)
+
+// clientLuksKeyStore adapts sdk.Client to the executor.LuksKeyStore interface.
+type clientLuksKeyStore struct {
+	client *sdk.Client
+}
+
+func (s *clientLuksKeyStore) GetKey(ctx context.Context, actionID string) (string, error) {
+	return s.client.GetLuksKey(ctx, actionID)
+}
+
+func (s *clientLuksKeyStore) StoreKey(ctx context.Context, actionID, devicePath, passphrase, reason string) error {
+	return s.client.StoreLuksKey(ctx, actionID, devicePath, passphrase, reason)
+}
